Add SSAA with SSIM-based perceptual downscaling
Super-sampling anti-aliasing with a conventional downscaling algorithm
smoothes the image, which is a disadvantage compared to disabled
anti-aliasing.
The SSIM-based perceptual downscaling algorithm tackles this problem and
it parallelises well on the GPU.

The new SSAA method can be enabled with the anti aliasing setting and it
is the slowest anti-aliasing technique now.