Do not discover mod directories that fail parsing
Fixes #15914

The root issue of the unit test failure is that all directories that are
found in the mod search are counted as mods, even if they are detected
to be invalid as such by the parser. For example, the presence of an
init.lua file is required, and the parser will return false if one is
not found. This return value was completely ignored. Simply counting the
mod conditionally on the parsing success makes the modserver tests pass
on MSVC. I have also added the nodiscard attribute to the parse function
as a preventive measure for regressions.