Avoid signal-unsafe operations in signal handler
TSan reported a signal-unsafe call to `ostream::put` in the POSIX signal
handler. This patch replaces the use of `dstream`, which generated those
calls, with calls to `write` and `strlen`, which are both async-signal
safe in POSIX.

This patch also changes the type of the global kill flag from `bool` to
`volatile std::sig_atomic_t`, to be certain there is no undefined
behavior. TSan did not mention that issue; I am not sure whether that is
something TSan is able to detect.