Adds back the missing endElement in parseTag for <action>
Before this change, <action> tags did not end until a space was found
or another tag was reached. Adding that call solves the issue, and with
some bold guess and seeing how spaces trigger ParsedText::ELEMENT_SEPARATOR,
this patch should be the way to fix this issue.

Fix https://discord.com/channels/369122544273588224/369123175583186964/1405238765647167600